#135: UI folder pages organizational refactor
