+++ lib/ui/pages/settings/account/change_password_page.dart
@@ -5,7 +5,7 @@ import 'package:noted_app/state/auth/auth_bloc.dart';
 import 'package:noted_app/state/auth/auth_event.dart';
 import 'package:noted_app/state/auth/auth_state.dart';
 import 'package:noted_app/ui/common/noted_library.dart';
-import 'package:noted_app/ui/settings/account/account_frame.dart';
+import 'package:noted_app/ui/pages/settings/account/account_frame.dart';
 import 'package:noted_app/util/extensions.dart';
 import 'package:noted_app/util/errors/noted_exception.dart';
