+++ lib/ui/pages/settings/settings_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:noted_app/ui/common/noted_library.dart';
-import 'package:noted_app/ui/settings/settings_row.dart';
+import 'package:noted_app/ui/pages/settings/settings_row.dart';
 import 'package:noted_app/util/extensions.dart';
 import 'package:noted_app/ui/router/noted_router.dart';
 import 'package:package_info_plus/package_info_plus.dart';
