+++ lib/ui/pages/settings/style/style_theme_page.dart
@@ -4,7 +4,7 @@ import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:noted_app/state/settings/settings_bloc.dart';
 import 'package:noted_app/state/settings/settings_event.dart';
 import 'package:noted_app/ui/common/noted_library.dart';
-import 'package:noted_app/ui/settings/style/style_frame.dart';
+import 'package:noted_app/ui/pages/settings/style/style_frame.dart';
 import 'package:noted_app/util/extensions.dart';
 import 'package:noted_models/noted_models.dart';
