+++ lib/ui/router/router_config.dart
@@ -2,17 +2,17 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:go_router/go_router.dart';
 import 'package:noted_app/state/auth/auth_bloc.dart';
 import 'package:noted_app/state/auth/auth_state.dart';
-import 'package:noted_app/ui/login/login_page.dart';
-import 'package:noted_app/ui/login/password_reset_page.dart';
-import 'package:noted_app/ui/login/register_page.dart';
-import 'package:noted_app/ui/login/sign_in_page.dart';
+import 'package:noted_app/ui/pages/login/login_page.dart';
+import 'package:noted_app/ui/pages/login/password_reset_page.dart';
+import 'package:noted_app/ui/pages/login/register_page.dart';
+import 'package:noted_app/ui/pages/login/sign_in_page.dart';
 import 'package:noted_app/ui/router/route_error_page.dart';
-import 'package:noted_app/ui/settings/account/account_page.dart';
-import 'package:noted_app/ui/settings/account/change_password_page.dart';
-import 'package:noted_app/ui/settings/settings_page.dart';
-import 'package:noted_app/ui/settings/style/style_page.dart.dart';
-import 'package:noted_app/ui/settings/style/style_fonts_page.dart';
-import 'package:noted_app/ui/settings/style/style_theme_page.dart';
+import 'package:noted_app/ui/pages/settings/account/account_page.dart';
+import 'package:noted_app/ui/pages/settings/account/change_password_page.dart';
+import 'package:noted_app/ui/pages/settings/settings_page.dart';
+import 'package:noted_app/ui/pages/settings/style/style_page.dart.dart';
+import 'package:noted_app/ui/pages/settings/style/style_fonts_page.dart';
+import 'package:noted_app/ui/pages/settings/style/style_theme_page.dart';
 import 'package:noted_app/ui/plugins/notebook/notebook_edit_page.dart';
 import 'package:noted_app/ui/plugins/notebook/notebook_page.dart';
