+++ noted_catalog/fonts/lora/Lora-Regular.ttf
