+++ noted_catalog/lib/catalog_settings.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:noted_app/ui/common/icon/noted_icons.dart';
 import 'package:noted_app/ui/common/layout/noted_header_page.dart';
-import 'package:noted_app/ui/settings/settings_row.dart';
-import 'package:noted_app/ui/settings/style/style_fonts_page.dart';
-import 'package:noted_app/ui/settings/style/style_theme_page.dart';
+import 'package:noted_app/ui/pages/settings/settings_row.dart';
+import 'package:noted_app/ui/pages/settings/style/style_fonts_page.dart';
+import 'package:noted_app/ui/pages/settings/style/style_theme_page.dart';
 
 class CatalogSettings extends StatelessWidget {
   const CatalogSettings({super.key});
